Buffer overflows are bad, so adjust the value of PlayerUpdatePLenMax to reflect the fact that timestamps have been changed from float to double.
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@18830 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ include/PlayerState.h
@@ -17,9 +17,9 @@
 #include "Address.h"
 #include "BufferedNetworkMessage.h"
 
-// 54 bytes
+// 58 bytes
 const int PlayerUpdatePLenMax =
-			sizeof(float)		+ // timestamp
+			sizeof(double)		+ // timestamp
 			PlayerIdPLen		+ // player id
 			sizeof(int32_t)	+ // order
 			sizeof(int16_t)	+ // status
